(PC-12131)[PRO]: EAC / offer form screen => rename category step into offer type following title
