--- pro/src/screens/OfferEducational/OfferEducationalForm/FormCategory/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FormCategory'
