+++ pro/src/screens/OfferEducational/OfferEducationalForm/FormOfferType/FormOfferType.tsx
