+++ pro/src/screens/OfferEducational/OfferEducationalForm/FormOfferType/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FormOfferType'
